Loede dev

+++ app/Http/Controllers/v1/WMS/Warehouse/WarehouseReceivingController.php
@@ -54,20 +54,6 @@ class WarehouseReceivingController extends Controller
             return $this->dataResponse('error', 400, $exception->getMessage());
         }
     }
-    #region Separate batches per item code
-    // public function onGetCurrent($referenceNumber, $status)
-    // {
-    //     $whereFields = [
-    //         'reference_number' => $referenceNumber,
-    //         'status' => $status // 0, 1
-    //     ];
-
-    //     $orderFields = [
-    //         'reference_number' => 'ASC'
-    //     ];
-    //     return $this->readCurrentRecord(WarehouseReceivingModel::class, null, $whereFields, null, $orderFields, null, 'Warehouse Receiving');
-    // }
-    #endregion
     public function onGetCurrent($referenceNumber, $status, $received_status = null)
     {
         try {
@@ -92,9 +78,6 @@ class WarehouseReceivingController extends Controller
 
             $isCompleteWarehouseReceive = WarehouseReceivingModel::select(
                 'reference_number',
-                DB::raw('SUM(substandard_quantity) as substandard_quantity'),
-                DB::raw('SUM(received_quantity) as received_quantity'),
-                DB::raw('SUM(JSON_LENGTH(produced_items)) as produced_items_count'),
                 DB::raw('(SUM(substandard_quantity) + SUM(received_quantity) = SUM(JSON_LENGTH(produced_items))) as is_completed')
             )
                 ->where('reference_number', $referenceNumber)
@@ -138,16 +121,21 @@ class WarehouseReceivingController extends Controller
             'action' => 'required|string|in:0,1', // 0 = Scan, 1 = Complete Transaction
         ]);
         try {
+            DB::beginTransaction();
             if ($fields['action'] == 0) {
                 $scannedItems = json_decode($fields['scanned_items'], true);
                 $this->onScanItems($scannedItems, $referenceNumber, $fields['created_by_id']);
+                $this->onCreatePutAway($scannedItems, $referenceNumber, $fields['created_by_id']);
+
             } else {
                 $this->onCompleteTransaction($referenceNumber, $fields['created_by_id']);
             }
 
+            DB::commit();
             return $this->dataResponse('success', 200, __('msg.update_success'));
 
         } catch (Exception $exception) {
+            DB::rollback();
             return $this->dataResponse('error', 400, 'Warehouse Receiving ' . $exception->getMessage());
         }
     }
@@ -175,8 +163,12 @@ class WarehouseReceivingController extends Controller
                         ->where('batch_number', $productionBatch->batch_number)
                         ->where('item_code', $itemCode)
                         ->first();
+
                     if ($warehouseReceiving) {
                         $warehouseForReceive = WarehouseForReceiveModel::where('reference_number', $referenceNumber)->update(['status' => 0]);
+                        $warehouseProducedItems = json_decode($warehouseReceiving->produced_items, true);
+                        $warehouseProducedItems[$itemDetails['sticker_no']]['status'] = '2.1';
+                        $warehouseReceiving->produced_items = json_encode($warehouseProducedItems);
                         $warehouseReceiving->received_quantity = ++$warehouseReceiving->received_quantity;
                         $warehouseReceiving->updated_by_id = $createdById;
                         $warehouseReceiving->save();
@@ -200,43 +192,39 @@ class WarehouseReceivingController extends Controller
                 ->get();
 
             if (count($warehouseReceiving) <= 0) {
-                throw new Exception('Warehouse Receiving reference number already received');
+                throw new Exception('Reference number already received');
             }
             $warehouseForReceiveItems = WarehouseForReceiveModel::where('reference_number', $referenceNumber)
                 ->where('created_by_id', $createdById)
                 ->orderBy('id', 'DESC')
                 ->first();
-            $receiveItemsArr = json_decode($warehouseForReceiveItems->production_items, true);
-
+            $receiveItemsArr = json_decode($warehouseForReceiveItems->production_items, true) ?? [];
+            if (count($receiveItemsArr) <= 0) {
+                throw new Exception('There are no items to be received from this repository');
+            }
             DB::beginTransaction();
             foreach ($warehouseReceiving as &$warehouseReceivingValue) {
                 $warehouseReceivingCurrentItemCode = $warehouseReceivingValue['item_code'];
                 $warehouseProducedItems = json_decode($warehouseReceivingValue['produced_items'], true);
                 $productionItemModel = $warehouseReceivingValue->productionBatch->productionItems;
-                $producedItems = json_decode($productionItemModel->produced_items, true);
+                // $producedItems = json_decode($productionItemModel->produced_items, true);
 
-
+                $discrepancy = [];
                 foreach ($warehouseProducedItems as $innerWarehouseReceivingKey => &$innerWarehouseReceivingValue) {
                     $flag = $this->onCheckItemReceive($receiveItemsArr, $innerWarehouseReceivingKey, $innerWarehouseReceivingValue, $warehouseReceivingCurrentItemCode);
-
-                    if ($flag) {
-                        $innerWarehouseReceivingValue['status'] = 3; // For Warehouse Receiving
-                        $producedItems[$innerWarehouseReceivingKey]['status'] = 3; // For Production Items
-                        $this->createProductionLog(ProductionItemModel::class, $productionItemModel->id, $producedItems[$innerWarehouseReceivingKey], $createdById, 1, $innerWarehouseReceivingKey);
+                    if (!$flag) {
+                        $innerWarehouseReceivingValue['sticker_no'] = $innerWarehouseReceivingKey;
+                        $discrepancy[] = $innerWarehouseReceivingValue;
                     }
                     unset($innerWarehouseReceivingValue);
                 }
-                $productionItemModel->status = 3;
-                $productionItemModel->save();
                 $warehouseForReceive = WarehouseForReceiveModel::where('reference_number', $referenceNumber)->delete();
-
                 $warehouseReceivingValue->status = 1;
                 $warehouseReceivingValue->updated_by_id = $createdById;
-                $warehouseReceivingValue->produced_items = json_encode($warehouseProducedItems);
+                $warehouseReceivingValue->discrepancy_data = json_encode($discrepancy);
                 $warehouseReceivingValue->save();
                 $this->createWarehouseLog(ProductionItemModel::class, $productionItemModel->id, WarehouseReceivingModel::class, $warehouseReceivingValue->id, $warehouseReceivingValue->getAttributes(), $createdById, 1);
             }
-
             DB::commit();
         } catch (Exception $exception) {
             DB::rollBack();
@@ -290,17 +278,15 @@ class WarehouseReceivingController extends Controller
                 $inclusionArray = [2];
                 $flag = $this->onItemCheckHoldInactiveDone(json_decode($productionItem->produced_items, true), $itemDetails['sticker_no'], $inclusionArray, []);
                 if ($flag) {
-                    $decodeItems = json_decode($productionItem->produced_items, true);
-                    $decodeItems[$itemDetails['sticker_no']]['status'] = 1.1;
-                    $productionItem->produced_items = json_encode($decodeItems);
-                    $productionItem->save();
-
                     $warehouseReceiving = WarehouseReceivingModel::where('reference_number', $referenceNumber)
                         ->where('production_order_id', $productionBatch->production_order_id)
                         ->where('batch_number', $productionBatch->batch_number)
                         ->where('item_code', $itemCode)
                         ->first();
                     if ($warehouseReceiving) {
+                        $warehouseReceivingProducedItems = json_decode($warehouseReceiving->produced_items, true);
+                        $warehouseReceivingProducedItems[$itemDetails['sticker_no']]['status'] = 1.1;
+                        $warehouseReceiving->produced_items = json_encode($warehouseReceivingProducedItems);
                         $warehouseReceiving->substandard_quantity = ++$warehouseReceiving->substandard_quantity;
                         $warehouseReceiving->save();
                     }
@@ -326,6 +312,70 @@ class WarehouseReceivingController extends Controller
         }
     }
 
+    public function onCreatePutAway($scannedItems, $referenceNumber, $createdById)
+    {
+        try {
+            $itemCodeArr = [];
+            $subLocationId = null;
+            foreach ($scannedItems as $value) {
+                $productionBatch = ProductionBatchModel::find($value['bid']);
+                $itemCode = $productionBatch->productionOta->itemMasterdata->item_code ?? $productionBatch->productionOtb->itemMasterdata->item_code;
+                if (!in_array($itemCode, $itemCodeArr)) {
+                    $itemCodeArr[] = $itemCode;
+                }
+                $productionItem = json_decode($productionBatch->productionItems->produced_items, true);
+                $subLocationId = $productionItem[$value['sticker_no']]['sub_location']['sub_location_id'];
+            }
+            $warehouseReceivingArr = [];
+            foreach ($itemCodeArr as $itemCode) {
+                $warehouseReceivingModel = WarehouseReceivingModel::select([
+                    'reference_number',
+                    'item_code',
+                    DB::raw('SUM(received_quantity) as received_quantity'),
+                ])
+                    ->where('reference_number', $referenceNumber)
+                    ->where('item_code', $itemCode)
+                    ->groupBy([
+                        'reference_number',
+                        'item_code'
+                    ])
+                    ->first();
+
+                if (!array_key_exists($itemCode, $warehouseReceivingArr)) {
+                    $warehouseReceivingArr[$itemCode] = $warehouseReceivingModel->getAttributes();
+                }
+            }
+
+            foreach ($warehouseReceivingArr as $warehouseReceiving) {
+                $warehouseReceivingModel = WarehouseReceivingModel::where('reference_number', $warehouseReceiving['reference_number'])
+                    ->where('item_code', $warehouseReceiving['item_code'])
+                    ->pluck('produced_items');
+                $mergedProducedItemsContainer = [];
+                foreach ($warehouseReceivingModel as $productionItems) {
+                    $items = json_decode($productionItems, true);
+                    $filteredItems = array_filter($items, function ($item) {
+                        return isset($item['status']) && $item['status'] == '2.1';
+                    });
+                    $mergedProducedItemsContainer = array_merge($mergedProducedItemsContainer, $filteredItems);
+                }
+
+                $warehousePutAwayController = new WarehousePutAwayController();
+                $warehousePutAwayRequest = new Request([
+                    'created_by_id' => $createdById,
+                    'warehouse_receiving_reference_number' => $warehouseReceiving['reference_number'],
+                    'received_quantity' => $warehouseReceiving['received_quantity'],
+                    'production_items' => json_encode($mergedProducedItemsContainer),
+                    'item_code' => $warehouseReceiving['item_code'],
+                    'scanned_items' => json_encode($scannedItems),
+                    'temporary_storage_id' => $subLocationId
+                ]);
+                $warehousePutAwayController->onCreate($warehousePutAwayRequest);
+            }
+        } catch (Exception $exception) {
+            throw new Exception($exception->getMessage());
+        }
+
+    }
 
     public function onCompleteTransactionMVP(Request $request, $reference_number)
     {
@@ -354,8 +404,8 @@ class WarehouseReceivingController extends Controller
 
                 // $queueTemporaryStorage
                 foreach ($itemsInQueue as $queuedItems) {
-                    $subLocationId = $queuedItems['sub_location']['sub_location_id'];
-                    $queuedTemporaryStorage = QueuedTemporaryStorageModel::find($subLocationId);
+                    $subLocationId = $queuedItems['sub_location']['sub_location_id'] ?? null;
+                    $queuedTemporaryStorage = QueuedTemporaryStorageModel::where('sub_location_id', $subLocationId)->first();
 
                     if ($queuedTemporaryStorage) {
                         $queuedTemporaryStorage->delete();
